Move patch logic to editfrom
